chore: bump ws
ws is a transitive dependency via socket.io and jsdom. There is a
security vulnerability (CVE-2024-37890) affecting <8.17.1 causing
DoS when handling a request with many HTTP headers.

I was able to remove our dependence on jsdom (keeping the types
definition for browser APIs used in the client).  The vulnerability
was handled in socket.io so I upgraded that as well.